django-env/requirements.txt
Django==1.7.4
Markdown==2.5.2
MySQL-python==1.2.5
Pillow==2.6.1
PyExecJS==1.1.0
PyReact==0.5.2
Pygments==2.0.2
dj-database-url==0.3.0
dj-static==0.0.6
django-bootstrap3==5.0.3
django-braces==1.4.0
django-filter==0.9.1
djangorestframework==3.0.3
httpie==0.8.0
httplib2==0.9
oauth2==1.5.211
oauthlib==0.7.2
requests==2.5.1
shortuuid==0.4.2
six==1.9.0
static3==0.5.1
wsgiref==0.1.2
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4
